+++ src/main/java/com/feiniaojin/ddd/ecosystem/pie/BootStrap.java
@@ -1,8 +1,5 @@
 package com.feiniaojin.ddd.ecosystem.pie;
 
-import com.feiniaojin.ddd.ecosystem.pie.Channel.Channel;
-import com.feiniaojin.ddd.ecosystem.pie.ChannelHandler.ChannelHandler;
-
 /**
  * 启动配置类
  *
