--- src/main/java/com/feiniaojin/ddd/ecosystem/pie/Channel/impl/DefaultChannel.java
@@ -1,10 +1,0 @@
-package com.feiniaojin.ddd.ecosystem.pie.Channel.impl;
-
-import com.feiniaojin.ddd.ecosystem.pie.AbstractChannel;
-
-/**
- * 默认通道实现
- *
- */
-public class DefaultChannel extends AbstractChannel {
-}
