+++ src/main/java/com/feiniaojin/ddd/ecosystem/pie/ChannelHandlerContext.java
@@ -1,8 +1,4 @@
-package com.feiniaojin.ddd.ecosystem.pie.ChannelHandlerContext;
-
-import com.feiniaojin.ddd.ecosystem.pie.Channel.Channel;
-import com.feiniaojin.ddd.ecosystem.pie.ChannelHandler.ChannelHandler;
-import com.feiniaojin.ddd.ecosystem.pie.ChannelPipeline.ChannelPipeline;
+package com.feiniaojin.ddd.ecosystem.pie;
 
 /**
  * ChannelHandlerContext
