+++ src/main/java/com/feiniaojin/ddd/ecosystem/pie/DefaultChannel.java
@@ -1,0 +1,8 @@
+package com.feiniaojin.ddd.ecosystem.pie;
+
+/**
+ * 默认通道实现
+ *
+ */
+public class DefaultChannel extends AbstractChannel {
+}
