+++ src/main/java/com/feiniaojin/ddd/ecosystem/pie/DefaultChannelHandlerContext.java
@@ -1,0 +1,20 @@
+package com.feiniaojin.ddd.ecosystem.pie;
+
+/**
+ * DefaultChannelHandlerContext默认实现
+ *
+ */
+public class DefaultChannelHandlerContext extends AbstractChannelHandlerContext {
+
+    private final ChannelHandler handler;
+
+    DefaultChannelHandlerContext(DefaultChannelPipeline pipeline, String name, ChannelHandler handler) {
+        super(pipeline, name, handler.getClass());
+        this.handler = handler;
+    }
+
+    @Override
+    public ChannelHandler handler() {
+        return handler;
+    }
+}
